pyproject.toml
[project]
name = "multi_agent"
version = "0.1.0"
description = "muluti agent using langgraph supervisor sample"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-anthropic>=0.3.10",
    "langchain-openai>=0.3.23",
    "langchain-core>=0.3.48",
    "langgraph-checkpoint>=2.0.21",
    "langgraph>=0.3.16",
    "langgraph-supervisor>=0.0.29",
    "mcp[cli]>=1.4.1",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
    "langfuse>=3.0.1",
    "langchain-community>=0.3.26",
    "streamlit>=1.46",
    "streamlit-cognito-auth>=1.3.1",
    "langchain-tavily",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langmem>=0.0.29",
    "psycopg2>=2.9.10",
    "asyncpg>=0.30.0",
    "langchain-mcp-adapters>=0.1.9",
    "fastmcp",
    "yfinance",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.1.77",
]
